Refactor handling of API errors (#287)
`halt` can only be used within the blocks and it will not work
in other places. Reusing code was very difficult since all the logic
should be placed in the ReST route blocks.

Now, API exception is introduced. This allows to move the common
functionalities to one place and handle this exception in ReST routes.

Signed-off-by: Aravinda Vishwanathapura <aravinda@kadalu.tech>